+++ tests/unit/Lookup/EntityRedirectApiLookupTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Addwiki\Wikibase\Tests\Unit\Api\Lookup;
+namespace Addwiki\Wikibase\Api\Tests\Unit\Lookup;
 
 use Addwiki\Mediawiki\Api\Client\Action\ActionApi;
 use Addwiki\Wikibase\Api\Lookup\EntityRedirectApiLookup;